k8s/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rusticai-k8s"
version = "0.1.0"
description = "Kubernetes-native execution engine for Rustic AI"
authors = ["Dragonscale Industries Inc. <dev@dragonscale.ai>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://www.rustic.ai/"
repository = "https://github.com/rustic-ai/rustic-ai"
packages = [{ include = "rustic_ai", from = "src" }]

[tool.poetry.urls]
"Rustic AI Core" = "https://pypi.org/project/rusticai-core/"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
rusticai-core = { path = "../core", develop = true }
# gRPC dependencies
grpcio = "^1.68.1"
grpcio-tools = "^1.68.1"
# K8s client
kubernetes = "^31.0.0"
# Redis client
redis = "^5.2.1"

[tool.poetry-monorepo.deps]


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = { version = "^24.2.0", extras = ["jupyter"] }
flake8 = "^7.1.2"
tox = "^4.24.1"
isort = "^6.0.0"
mypy = "^1.15.0"
flaky = "^3.8.1"
pytest-asyncio = "^0.26.0"
fakeredis = "^2.27.0"
rusticai-testing = {path = "../testing", develop = true}

[tool.poetry.extras]
test = ["pytest"]

[tool.poetry.scripts]
rustic-agent-host = "rustic_ai.k8s.agent_host.server:serve"

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = true
plugins = "pydantic.mypy"
explicit_package_bases = true


[tool.isort]
profile = "black"
force_sort_within_sections = true # For the sorting order issues
lines_between_sections = 1 # For the newline issue
known_third_party = ["pydantic"]
known_first_party = ["rustic_ai.core"]
